requirements.txt
pinecone-client
pdfplumber
tqdm
sentence-transformers==2.7.0
flask
pypdf2
transformers
llama-index
llama_index.embeddings.huggingface
llama_index.llms.huggingface==0.3.5
llama_index.llms.huggingface_api
llama-index-embeddings-instructor==0.2.1
llama_index.vector_stores.pinecone
llama_index.core
ipython
nltk
torch==2.4.1
accelerate==1.1.0
streamlit
